Moved numpy utils into services/mlmodel/utils.py; changed imports accordingly; made tests have less irrelevant warnings
